requirements.txt
google-genai>=0.7.0
discord.py>=2.3.0
schedule>=1.2.0
requests>=2.31.0
flask>=2.3.0
python-dotenv>=1.0.0
